Add User getOwn article api, get article list api, get targer article detail.

+++ src/route/article/index.ts
@@ -1,11 +1,11 @@
 import { ArticleModel } from "@codesRoot/db/schemas/article";
 import { ArticleDocument } from "@codesRoot/db/schemas/articleType";
-import { UserModel } from "@codesRoot/db/schemas/user";
 import { UserDocument } from "@codesRoot/db/schemas/userType";
-import { returnIfNotPass, uploadArticleCheck } from "@codesRoot/db/schemas/validator";
+import { articleIdCheck, articleQueryCheck, returnIfNotPass, uploadArticleCheck } from "@codesRoot/db/schemas/validator";
 import { IFoundError } from "@codesRoot/middleware/errorType";
 import { multerconfig, processImagetojpg } from "@codesRoot/middleware/imageConfig";
 import { RequestLogin } from "@codesRoot/middleware/LoginRequest";
+import { GetLatestArticleList } from "@codesRoot/utils/helper";
 import { NextFunction, Request, Response, Router } from "express";
 import { matchedData } from "express-validator";
 
@@ -37,58 +37,72 @@ articleRouter.post('/upload',
             await user.updateOne({ $push: { ArticleIDList: newArticle.id } })
             return res.status(200).json({ message: "Article Upload.", id: newArticle.id })
         } catch (err) {
-
-            if (err instanceof Error) {
-                if (err.name == "MongoServerError") {
-                    return next(new IFoundError(err.name, 409, [err.message]));
-                }
-
-                return next(err);
-
-            }
+            return next(err);
         }
     })
 
-articleRouter.get("/uploadCount",
-    RequestLogin,
-    (req: Request, res: Response) => {
-        const user = req.user as UserDocument;
-        return res.status(200).json({ count: user.ArticleIDList.length })
-    })
 
-
-articleRouter.delete("/:id",
-    RequestLogin,
+articleRouter.get("/:id",
+    articleIdCheck,
+    returnIfNotPass,
     async (req: Request, res: Response, next: NextFunction) => {
         try {
-            const user = req.user as UserDocument;
-            const findArticle = await ArticleModel.findById(req.params.id);
+            const data = matchedData(req);
+            const findArticle = await ArticleModel.findById(data.id);
             if (!findArticle) {
                 return next(new IFoundError("Article Not Found.", 404, [`Article ID: ${req.params.id} not Found`]))
+            }
+            const articleData = {
+                id: findArticle.id,
+                Image: process.env.IMAGEPATH + findArticle.Image,
+                Tags: findArticle.Tags,
+                Name: findArticle.Name,
+                DetailInfo: findArticle.DetailInfo,
+                OwnerID: findArticle.OwnerID,
+                CreatedAt: findArticle.CreatedAt,
+                UpdatedAt: findArticle.UpdatedAt,
+            }
+            return res.status(200).json(articleData)
+        } catch (err) {
+            return next(err);
+        }
+    })
+articleRouter.get("/",
+    articleQueryCheck,
+    async (req: Request, res: Response, next: NextFunction) => {
+        try {
+            const data = matchedData(req);
+            data.From = data.From === undefined ? 0 : data.From
+            data.Length = data.Length === undefined ? 30 : data.Length
+            const articleList: Array<ArticleDocument> = await ArticleModel.find();
+            const articleListMap = GetLatestArticleList(articleList, data.From, data.Length);
+            res.status(200).json({ ArticleList: articleListMap })
+        } catch (err) {
+            next(err)
+        }
+    })
+articleRouter.delete("/:id",
+    RequestLogin,
+    articleIdCheck,
+    returnIfNotPass,
+    async (req: Request, res: Response, next: NextFunction) => {
+        try {
+
+            const data = matchedData(req);
+            const user = req.user as UserDocument;
+            const findArticle = await ArticleModel.findById(data.id);
+            if (!findArticle) {
+                return next(new IFoundError("Article Not Found.", 404, [`Article ID: ${data.id} not Found`]))
             }
             const isOwn = findArticle.OwnerID.equals(user.id)
             if (!isOwn && user.Permission !== 1) {
                 return next(new IFoundError("Insufficient permissions", 403, ["This Article not belong User."]))
             }
-            if (isOwn) {
-                user.updateOne({ $pull: { ArticleIDList: req.params.id } })
-            }
-            else {
-                await UserModel.updateOne({
-                    ArticleIDList: req.params.id
-                }, {
-                    $pull: { ArticleIDList: req.params.id }
-                })
-            }
+            const ownUser = await findArticle.getOwner()
+            await ownUser.updateOne({ $pull: { ArticleIDList: data.id } })
             await ArticleModel.deleteOne({ _id: findArticle._id })
             return res.sendStatus(200)
         } catch (err) {
-            if (err instanceof Error) {
-                if (err.name == "MongoServerError") {
-                    return next(new IFoundError(err.name, 409, [err.message]));
-                }
-            }
-
             return next(err);
         }
     })
